requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==1.16.0
Flask==2.2.5
gitdb==4.0.12
GitPython==3.1.45
idna==2.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==2.1.3
narwhals==2.0.1
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
pyarrow==21.0.0
pydeck==0.9.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.31.0
requests-oauthlib==1.2.0
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.18
watchdog==6.0.0
websocket-client==1.6.1
Werkzeug==2.2.3

